requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
pydantic==2.9.2
Jinja2==3.1.4
httpx==0.27.2
python-multipart==0.0.9
itsdangerous==2.2.0 